+++ src/pages/Lend/LoanChannelCategoryControl.vue
@@ -7,6 +7,7 @@
 					v-else
 					:to="filterUrl"
 					class="tw-text-action tw-flex tw-items-center tw-float-right"
+					@click.native="trackAdvancedFilters"
 				>
 					<img class="tw-w-2 tw-mr-1" src="@/assets/images/tune.svg">
 					Advanced filters
@@ -62,6 +63,7 @@
 						:remaining-loans="helpmeChooseRemainingLoans"
 						:items-in-basket="itemsInBasket"
 						:is-visitor="isVisitor"
+						:user-data="userData"
 					/>
 				</div>
 				<div v-else class="loan-card-group row small-up-1 large-up-2 xxlarge-up-3">
@@ -92,6 +94,7 @@
 						:remaining-loans="helpmeChooseRemainingLoans"
 						:items-in-basket="itemsInBasket"
 						:is-visitor="isVisitor"
+						:user-data="userData"
 					/>
 				</div>
 				<kv-pagination
@@ -250,7 +253,8 @@ export default {
 				}]
 			},
 			filtersLoaded: false,
-			selectedQuickFilters: {}
+			selectedQuickFilters: {},
+			userData: {}
 		};
 	},
 	computed: {
@@ -406,6 +410,7 @@ export default {
 
 		// Set user status
 		this.isVisitor = !_get(pageQueryData, 'my.userAccount.id');
+		this.userData = _get(pageQueryData, 'my.userAccount');
 
 		// Filter routes on param.category to get current path
 		this.targetedLoanChannelURL = _get(this.$route, 'params.category');
@@ -469,6 +474,13 @@ export default {
 		}
 	},
 	methods: {
+		trackAdvancedFilters() {
+			this.$kvTrackEvent(
+				'Search',
+				'click',
+				'category-advanced-filters'
+			);
+		},
 		resetFilters() {
 			this.selectedQuickFilters = {};
 		},
@@ -657,6 +669,8 @@ export default {
 		quickFiltersFlssParameters(matchedUrls = []) {
 			if (this.targetedLoanChannelURL === 'single-parents') {
 				this.flssLoanSearch = { tagId: [17] };
+			} else if (this.targetedLoanChannelURL === 'livestock') {
+				this.flssLoanSearch = { activityId: [73] };
 			} else {
 				this.flssLoanSearch = matchedUrls[0]?.flssLoanSearch ?? {};
 			}