+++ src/pages/Thanks/ThanksPage.vue
@@ -125,6 +125,7 @@ import orderBy from 'lodash/orderBy';
 import thanksPageQuery from '@/graphql/query/thanksPage.graphql';
 import { processPageContentFlat } from '@/util/contentfulUtils';
 import { userHasLentBefore, userHasDepositBefore } from '@/util/optimizelyUserMetrics';
+import setHotJarUserAttributes from '@/util/hotJarUserAttributes';
 import logFormatter from '@/util/logFormatter';
 import { joinArray } from '@/util/joinArray';
 import KvButton from '~/@kiva/kv-components/vue/KvButton';
@@ -264,6 +265,7 @@ export default {
 			logReadQueryError(e, `Thanks page readQuery failed: (transaction_id: ${transactionId})`);
 		}
 
+		const hasEverLoggedIn = data?.hasEverLoggedIn;
 		const modernSubscriptions = data?.mySubscriptions?.values ?? [];
 		this.hasModernSub = modernSubscriptions.length !== 0;
 		this.lender = {
@@ -301,6 +303,14 @@ export default {
 
 		userHasLentBefore(hasLentBefore);
 		userHasDepositBefore(hasDepositBefore);
+
+		// MARS-246 Hotjar user attributes
+		setHotJarUserAttributes({
+			userId: data?.my?.userAccountId?.id,
+			hasEverLoggedIn,
+			hasLentBefore,
+			hasDepositBefore,
+		});
 
 		if (!this.isGuest && !data?.my?.userAccount) {
 			logFormatter(